+++ frontend/src/pages/AboutUs/AboutUsPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function AboutUsPage() {
+  return <div>Placeholder</div>;
+}
+
+export default AboutUsPage;
