+++ frontend/src/pages/DataCatalog/dataset_explorer/DatasetExplorer.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables";
+@import "../../../styles/variables";
 
 .DatasetExplorer {
   display: flex;
