+++ frontend/src/pages/DataCatalog/dataset_explorer/DatasetListing.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables";
+@import "../../../styles/variables";
 
 .FlexRow {
   display: flex;
