+++ frontend/src/pages/Landing/LandingPage.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables";
+@import "../../styles/variables";
 
 .Grid {
   margin: auto;
